refactor: move orchestration files to proxy/orchestration/ subdirectory
- orchestrator, resilience, semaphore, scope, retry-rules -> proxy/orchestration/
- Fixed all import paths in proxy/, admin/, core/, db/, tests/
- Broke circular import: retry-rules.ts now imports from db/retry-rules.js directly
- All 580 tests pass
