+++ src/index.ts
@@ -20,8 +20,8 @@ import { authMiddleware } from "./middleware/auth.js";
 import { openaiProxy } from "./proxy/openai.js";
 import { anthropicProxy } from "./proxy/anthropic.js";
 import { adminRoutes } from "./admin/routes.js";
-import { RetryRuleMatcher } from "./proxy/retry-rules.js";
-import { ProviderSemaphoreManager } from "./proxy/semaphore.js";
+import { RetryRuleMatcher } from "./proxy/orchestration/retry-rules.js";
+import { ProviderSemaphoreManager } from "./proxy/orchestration/semaphore.js";
 import { AdaptiveConcurrencyController } from "./proxy/adaptive-controller.js";
 import { loadEnhancementConfig } from "./proxy/enhancement-config.js";
 import type { StateRegistry } from "./core/registry.js";
