+++ src/proxy/orchestration/orchestrator.ts
@@ -1,7 +1,7 @@
 import type { FastifyReply, FastifyRequest } from "fastify";
-import type { TransportResult } from "./types.js";
-import { ProviderSwitchNeeded } from "./types.js";
-import type { Target, ConcurrencyOverride } from "../core/types.js";
+import type { TransportResult } from "../types.js";
+import { ProviderSwitchNeeded } from "../types.js";
+import type { Target, ConcurrencyOverride } from "../../core/types.js";
 import type { ResilienceLayer, ResilienceResult, ResilienceConfig } from "./resilience.js";
 import { ResilienceLayer as ResilienceLayerClass } from "./resilience.js";
 import type { RetryRuleMatcher } from "./retry-rules.js";
@@ -9,10 +9,10 @@ import type { SemaphoreScope } from "./scope.js";
 import { SemaphoreScope as SemaphoreScopeClass } from "./scope.js";
 import type { TrackerScope } from "./scope.js";
 import { TrackerScope as TrackerScopeClass } from "./scope.js";
-import type { ActiveRequest } from "../monitor/types.js";
+import type { ActiveRequest } from "../../monitor/types.js";
 import type { ProviderSemaphoreManager } from "./semaphore.js";
-import type { RequestTracker } from "../monitor/request-tracker.js";
-import type { AdaptiveConcurrencyController } from "./adaptive-controller.js";
+import type { RequestTracker } from "../../monitor/request-tracker.js";
+import type { AdaptiveConcurrencyController } from "../adaptive-controller.js";
 
 const DEFAULT_BASE_DELAY_MS = 1000;
 const DEFAULT_FAILOVER_THRESHOLD = 400;
