+++ src/proxy/orchestration/resilience.ts
@@ -1,8 +1,8 @@
-import { MS_PER_SECOND } from "../core/constants.js";
+import { MS_PER_SECOND } from "../../core/constants.js";
 import type { RetryRuleMatcher } from "./retry-rules.js";
-import { ProviderSwitchNeeded } from "./types.js";
-import type { TransportResult } from "./types.js";
-import type { Target } from "../core/types.js";
+import { ProviderSwitchNeeded } from "../types.js";
+import type { TransportResult } from "../types.js";
+import type { Target } from "../../core/types.js";
 
 // ---------- Strategy Pattern ----------
