+++ tests/anthropic-proxy.test.ts
@@ -6,7 +6,7 @@ import { encrypt } from "../src/utils/crypto.js";
 import { anthropicProxy } from "../src/proxy/anthropic.js";
 import { initDatabase } from "../src/db/index.js";
 import { setSetting } from "../src/db/settings.js";
-import { ProviderSemaphoreManager } from "../src/proxy/semaphore.js";
+import { ProviderSemaphoreManager } from "../src/proxy/orchestration/semaphore.js";
 import { RequestTracker } from "../src/monitor/request-tracker.js";
 
 const TEST_ENCRYPTION_KEY =
