+++ tests/monitor/request-tracker.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeEach, afterEach, vi, type Mock } from "vitest";
 import type { ServerResponse } from "node:http";
 import { RequestTracker } from "../../src/monitor/request-tracker.js";
-import { ProviderSemaphoreManager } from "../../src/proxy/semaphore.js";
+import { ProviderSemaphoreManager } from "../../src/proxy/orchestration/semaphore.js";
 import type { ActiveRequest } from "../../src/monitor/types.js";
 
 // --- Mocks ---
